% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.optimal.CS.R
\name{approx.optimal.CS}
\alias{approx.optimal.CS}
\title{combines result from subsamples using the approximate optimal 
weights together with their proper variances}
\usage{
approx.optimal.CS(nk, ck, mu.split.est, sigma2.split.est, d.split.est,
  sigma2.split.var, d.split.var)
}
\arguments{
\item{nk}{a vector containing the cluster sizes}

\item{ck}{a vector containing the number of clusters of size nk}

\item{mu.split.est}{a vector containing the \eqn{\widehat{\mu}}'s from all sub-samples.}

\item{sigma2.split.est}{a vector containing the \eqn{\widehat{\sigma}^2}'s from all sub-samples}

\item{d.split.est}{a vector containing the \eqn{\widehat{d}}'s from all sub-samples}

\item{sigma2.split.var}{a vector containing the \eqn{\mathrm{Var}(\widehat{\sigma}^2)}'s from all sub-samples}

\item{d.split.var}{a vector containing the \eqn{\mathrm{Var}(\widehat{d})}'s from all sub-samples}
}
\value{
a list with computed materials
}
\description{
combines result from subsamples using the approximate optimal 
weights together with their proper variances
}
\author{
Vahid Nassiri
}
