% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBYcluster.CS.R
\name{clusterBYcluster.CS}
\alias{clusterBYcluster.CS}
\title{computes weighted, two stage, unbiased two stage methods for cluster by cluster analyses}
\usage{
clusterBYcluster.CS(nk, ck, Data)
}
\arguments{
\item{nk}{a vector containing the cluster sizes}

\item{ck}{a vector containing the number of clusters of size nk}

\item{Data}{a 3-column matrix with first column the subject, second 
column the response variable, and third column the split, which show which 
observation  belongs to which sub-sample}
}
\value{
a list with computed materials
}
\description{
computes weighted, two stage, unbiased two stage methods for cluster by cluster analyses
}
\author{
Vahid Nassiri
}
