% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.CS.all.R
\name{est.CS.all}
\alias{est.CS.all}
\title{estimates the parameters for all of the splits simultaneously}
\usage{
est.CS.all(data, ck, nk)
}
\arguments{
\item{data}{a 2-column matrix with response variable as its first column 
and splits indicator as its second column. All clusters should 
be of equal size by placing NaN to make smaller clusters the 
same size as the largest one.}

\item{ck}{a vector containing the number of clusters of size nk}

\item{nk}{a vector containing the cluster sizes}
}
\value{
a list with computed materials
}
\description{
estimates the parameters for all of the splits simultaneously
}
\author{
Vahid Nassiri
}
