% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.CS.for.R
\name{est.CS.for}
\alias{est.CS.for}
\title{computes the parameters estimates with their variances within each split}
\usage{
est.CS.for(n, C, Y)
}
\arguments{
\item{n}{an integer indicating the common cluster size within each split}

\item{C}{an integer indicating the number of clusters within each split}

\item{Y}{a vector containing the response values corresponding to each split}
}
\value{
a list with computed materials
}
\description{
computes the parameters estimates with their variances within each split
}
\author{
Vahid Nassiri
}
