% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.free.CS.R
\name{param.free.CS}
\alias{param.free.CS}
\title{computes the three parameter free combining rules: Prop., Equal, and Appr.sc.}
\usage{
param.free.CS(nk, ck, mu.split.est, sigma2.split.est, d.split.est,
  mu.split.var, sigma2.split.var, d.split.var)
}
\arguments{
\item{nk}{a vector containing the cluster sizes}

\item{ck}{a vector containing the number of clusters of size nk}

\item{mu.split.est}{a vector containing the \eqn{\widehat{\mu}}'s from all sub-samples.}

\item{sigma2.split.est}{a vector containing the \eqn{\widehat{\sigma}^2}'s from all sub-samples}

\item{d.split.est}{a vector containing the \eqn{\widehat{d}}'s from all sub-samples}

\item{mu.split.var}{a vector containing the variance of \eqn{\widehat{\mu}}'s from all sub-samples}

\item{sigma2.split.var}{a vector containing the \eqn{\mathrm{Var}(\widehat{\sigma}^2)}'s from all sub-samples}

\item{d.split.var}{a vector containing the \eqn{\mathrm{Var}(\widehat{d})}'s from all sub-samples}
}
\value{
a list with computed materials
}
\description{
computes the three parameter free combining rules: Prop., Equal, and Appr.sc.
}
\author{
Vahid Nassiri
}
